+++ src/main/java/org/codingtext/admin/controller/AuthController.java
@@ -1,9 +1,9 @@
 package org.codingtext.admin.controller;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+
 import org.codingtext.admin.dto.LoginRequest;
-import org.codingtext.admin.dto.PermitRequest;
+
 import org.codingtext.admin.dto.SignupRequest;
 import org.codingtext.admin.service.AuthService;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/auth")
-@Slf4j
 public class AuthController {
 
     private final AuthService authService;
@@ -25,17 +24,5 @@ public class AuthController {
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody LoginRequest loginRequest) {
         return ResponseEntity.ok(authService.login(loginRequest));
-    }
-
-    @GetMapping("/none")
-    public ResponseEntity<?> findNoneAccount() {
-        return ResponseEntity.ok(authService.findNoneAccount());
-    }
-
-    @PostMapping("/permit")
-    public ResponseEntity<?> permitAdmin(
-            @RequestHeader("AdminId") long adminId,
-            @RequestBody PermitRequest permitRequest) {
-        return ResponseEntity.ok(authService.processAdminRequest(adminId, permitRequest));
     }
 }
