+++ src/main/java/org/codingtext/admin/service/AuthService.java
@@ -10,8 +10,6 @@ import org.codingtext.admin.repository.AdminRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -64,46 +62,5 @@ public class AuthService {
                 .build();
 
         return new LoginResponse(new JwtToken(accessToken, null), adminResponse);
-    }
-
-    public List<AdminResponse> findNoneAccount() {
-        // DB에서 NONE 역할의 Admin을 조회하고 DTO로 변환
-        return adminRepository.findByAdminRole(AdminRole.NONE).stream()
-                .map(admin -> AdminResponse.builder()
-                        .id(admin.getId())
-                        .email(admin.getEmail())
-                        .adminRole(admin.getAdminRole()) // Enum을 문자열로 변환
-                        .build())
-                .collect(Collectors.toList());
-    }
-
-    public PermitResponse processAdminRequest(long adminId, PermitRequest permitRequest) {
-        // root 조회
-        Admin rootAdmin = adminRepository.findById(adminId)
-                .orElseThrow(() -> new RuntimeException("Root admin not found"));
-        // none 조회
-        Admin noneAdmin = adminRepository.findById(permitRequest.getAdminId())
-                .orElseThrow(() -> new RuntimeException("None admin not found"));
-        // ROLE이 ROOT인 경우에만 처리
-        if (rootAdmin.getAdminRole() == AdminRole.ROOT) {
-            if (permitRequest.getIsPermit()) {
-                // NONE 관리자 권한을 GENERAL로 변경
-                adminRepository.save(noneAdmin.toBuilder()
-                        .adminRole(AdminRole.GENERAL)
-                        .build());
-                return PermitResponse.builder()
-                        .adminId(permitRequest.getAdminId())
-                        .message("권한이 GENERAL로 변경되었습니다.")
-                        .build();
-            } else {
-                adminRepository.deleteById(permitRequest.getAdminId());
-                return PermitResponse.builder()
-                        .adminId(permitRequest.getAdminId())
-                        .message("해당 계정이 삭제되었습니다.")
-                        .build();
-            }
-        } else {
-            throw new RuntimeException("Only ROOT accounts can process this request.");
-        }
     }
 }
